.agent/skills/content-repurposing/requirements.txt
apify-client==1.6.4
yt-dlp
notion-client
openai
requests
google-generativeai
